requirements.txt
aiogram==2.*
motor
pymongo
python-dotenv
llama-cpp-python
huggingface-hub